Allow to capture regex group.
The expected stdout/stderr of a process can use regex
to match and capture some group. The group can then be
used to define commands or expected stdout/stderr.

That feature is used to remove some commands from
test-and-coverage.sh and replace them by an intake
workflow (well not yet a Workflow, just a sequence
of Job).